requirements.txt
openai>=1.0.0
pandas>=2.0.0
pyngrok>=7.0.0
requests>=2.32.5
python-dotenv>=1.2.1
